+++ src/main/java/org/example/utils/JwtUtil.java
@@ -11,7 +11,7 @@ public class JwtUtil {
     private static final int tokenLiveTime = 1000 * 3600 * 96; // 4-day
     private static final String secretKey = "very_long_mazgiskjdh2skjdhadasdasg7fgdfgdfdftrhdgrgefergetdgsfegvergdgsbdzsfbvgdsetbgrFLKWRMF.KJERNGVSFUOISN;IUVNSDBFIUSH;IULFHWA;UOIESIU;OF;IOEJ'OIGJMKLDFMGghjgjOTFIJBP";
 
-    public static String encode(Integer profileId,String username, ProfileRole role) {
+    public static String encode(Long profileId,String username, ProfileRole role) {
         JwtBuilder jwtBuilder = Jwts.builder();
         jwtBuilder.issuedAt(new Date());
